+++ demo/views/layout/row-demo/row-demo.js
@@ -34,11 +34,12 @@ class RowDemo extends React.Component {
    */
   get demo() {
     return (
-      <Row columns={ this.value('columns') }>
+      <Row columns={ this.value('columns') } gutter={ this.value('gutter') }>
         {
           this.value('columnData').map((data, index) => {
             return (
               <div
+                style={{ backgroundColor: '#efefef' }}
                 key={ index }
                 columnOffset={ data.get('columnOffset') }
                 columnSpan={ data.get('columnSpan') }
@@ -61,6 +62,10 @@ class RowDemo extends React.Component {
 
     if (this.value('columns')) {
       html += `\n  columns='${this.value('columns')}'\n`;
+    }
+
+    if (this.value('gutter')) {
+      html += `\n  gutter='${this.value('gutter')}'\n`;
     }
 
     html += ">\n";
@@ -188,6 +193,14 @@ class RowDemo extends React.Component {
             labelInline={ true }
             onChange={ this.action.bind(this, 'columns') }
             placeholder="Enter a fixed number of columns"
+          />
+
+          <Number
+            label="Gutter"
+            value={ this.value('gutter') }
+            labelInline={ true }
+            onChange={ this.action.bind(this, 'gutter') }
+            placeholder="Enter width in px for the gutter"
           />
         </Row>