+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ## Component Enhancements
 
 * `Link` component now has a prop of `iconAlign` to align icons to the right of the link's text.
+* `Row` component can now be given a custom value for the size of the gutter using the prop `gutter`.
 
 ## Minor Improvements